% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query.R
\name{create_query}
\alias{create_query}
\title{Create queries to the Riot API}
\usage{
create_query(..., param, version, region)
}
\arguments{
\item{...}{named key-value pairs}

\item{version}{which api version do you need}

\item{region}{region to run query for}
}
\value{
returns formatted url
}
\description{
Create queries to the Riot API
}
\examples{
\dontrun{NA}
}

