% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champion_mastery.R
\name{get_all_champ_mastery}
\alias{get_all_champ_mastery}
\title{Get all champion mastery entries sorted by number of champion points descending}
\usage{
get_all_champ_mastery(player_id, region = "NA1")
}
\arguments{
\item{player_id}{summoner id}

\item{region}{platform id}
}
\value{
all champion masteries
}
\description{
Get all champion mastery entries sorted by number of champion points descending
}
\examples{
get_all_champ_mastery(29400662)
}

