% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champion_mastery.R
\name{get_champ_mastery}
\alias{get_champ_mastery}
\title{Get a champion mastery by player id and champion id. Response code 204 means there were no masteries found for given player id or player id and champion id combination.}
\usage{
get_champ_mastery(player_id, champ_id, region = "NA1")
}
\arguments{
\item{player_id}{summoner id}

\item{champ_id}{champmion id}

\item{region}{platform id}
}
\value{
player's champion mastery on a single champ
}
\description{
Get a champion mastery by player id and champion id. Response code 204 means there were no masteries found for given player id or player id and champion id combination.
}
\examples{
get_champ_mastery(29400662, 23)
}

