% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champion_mastery.R
\name{get_champ_mastery_score}
\alias{get_champ_mastery_score}
\title{Get a player's total champion mastery score, which is sum of individual champion mastery levels}
\usage{
get_champ_mastery_score(player_id, region = "NA1")
}
\arguments{
\item{player_id}{summoner id}

\item{region}{platform id}
}
\value{
player's total champion mastery score
}
\description{
Get a player's total champion mastery score, which is sum of individual champion mastery levels
}
\examples{
get_champ_mastery_score(29400662)
}

