% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champion_mastery.R
\name{get_champ_mastery_top}
\alias{get_champ_mastery_top}
\title{Get specified number of top champion mastery entries sorted by number of champion points descending}
\usage{
get_champ_mastery_top(player_id, count = NULL, region = "NA1")
}
\arguments{
\item{player_id}{summoner id}

\item{count}{optional, max number of entries}

\item{region}{platform id}
}
\value{
top champion mastery entries
}
\description{
Get specified number of top champion mastery entries sorted by number of champion points descending
}
\examples{
get_champ_mastery_top(29400662, 1231241)
}

