% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champion.R
\name{get_champion}
\alias{get_champion}
\title{Get champion information}
\usage{
get_champion(id, region = "na")
}
\arguments{
\item{id}{champion id}
}
\value{
data frame of information about a champion
}
\description{
Get champion information
}
\examples{
get_champion("32")
}

