% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_game.R
\name{get_current_game}
\alias{get_current_game}
\title{Get current game information for the given summoner ID}
\usage{
get_current_game(platform_id, summoner_id, region = "na")
}
\arguments{
\item{platform_id}{Region to Execute Against}

\item{summoner_id}{The ID of the summoner.}
}
\value{
current game info for given summoner
}
\description{
Get current game information for the given summoner ID
}
\examples{

}

