% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featured_games.R
\name{get_featured_games}
\alias{get_featured_games}
\title{Get list of featured games.}
\usage{
get_featured_games(region = "na")
}
\arguments{
\item{region}{Region to Execute Against}
}
\value{
featured game info
}
\description{
Get list of featured games.
}
\examples{
get_featured_games()
}

