% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_id_from_name}
\alias{get_id_from_name}
\title{Get summoner ID from summoner name}
\usage{
get_id_from_name(summoner_name)
}
\arguments{
\item{summoner_name}{}
}
\description{
Get summoner ID from summoner name
}

