% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{get_league_by_summoner}
\alias{get_league_by_summoner}
\title{Get leagues mapped by summoner ID for a given list of summoner IDs}
\usage{
get_league_by_summoner(ids, region = "na")
}
\arguments{
\item{ids}{list of summoner id's}

\item{region}{region to run query for}
}
\value{
league information for given summoners
}
\description{
Get leagues mapped by summoner ID for a given list of summoner IDs
}
\examples{
get_league_by_summoner(c(29400662, 521955))
}

