% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{get_league_by_team}
\alias{get_league_by_team}
\title{Get league entries mapped by summoner ID for a given list of summoner IDs}
\usage{
get_league_by_team(ids, region = "na")
}
\arguments{
\item{ids}{list of summoner id's}

\item{region}{region to run query for}
}
\value{
league entries for given summoners
}
\description{
Get league entries mapped by summoner ID for a given list of summoner IDs
}
\examples{
get_league_by_team("TEAM-75796040-4c8c-11e6-9d8d-c81f66dcfb5a")
}

