% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{get_league_challenger}
\alias{get_league_challenger}
\title{Get challenger tier leagues.}
\usage{
get_league_challenger(type, region = "na")
}
\arguments{
\item{type}{one of "RANKED_SOLO_5x5", "RANKED_TEAM_5x5", "RANKED_TEAM_3x3"}

\item{region}{region to run query for}
}
\value{
data about challenger tier league in a queue
}
\description{
Get challenger tier leagues.
}
\examples{
get_league_challenger("RANKED_SOLO_5x5")
}

