% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{get_league_master}
\alias{get_league_master}
\title{Get master tier leagues.}
\usage{
get_league_master(type, region = "na")
}
\arguments{
\item{type}{one of "RANKED_SOLO_5x5", "RANKED_TEAM_5x5", "RANKED_TEAM_3x3"}

\item{region}{region to run query for}
}
\value{
data about master tier league in a queue
}
\description{
Get master tier leagues.
}
\examples{
get_league_master("RANKED_SOLO_5x5")
}

