% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{get_match}
\alias{get_match}
\title{Retrieve match by match ID}
\usage{
get_match(match, region)
}
\arguments{
\item{match}{id}

\item{region}{for api}
}
\value{
match information
}
\description{
Retrieve match by match ID
}
\examples{
get_match("2269165589", "na")
}

