% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{get_matchlist}
\alias{get_matchlist}
\title{Retrieve match list by summoner ID.}
\usage{
get_matchlist(summoner, region)
}
\arguments{
\item{summoner}{summoner id}

\item{region}{region id}
}
\value{
matchlist for a summoner
}
\description{
TODO: add filtering parameters
}
\details{
https://na.api.pvp.net/api/lol/na/v2.2/matchlist/by-summoner/5908?rankedQueues=TEAM_BUILDER_DRAFT_RANKED_5x5&seasons=SEASON2016&api_key=5f309460-6a23-4f50-8f07-3c6d90780a80
}
\examples{
get_matchlist(5908, "na")
}

