% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riot_key.R
\name{get_riot_key}
\alias{get_riot_key}
\title{Get the currently set Riot API key}
\usage{
get_riot_key()
}
\value{
if set, the Riot API key
}
\description{
Get the currently set Riot API key
}
\examples{
print(get_riot_key())
}

