% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_stats}
\alias{get_stats}
\title{Get ranked/summary stats by summoner ID}
\usage{
get_stats(summoner, region, type = "ranked")
}
\arguments{
\item{summoner}{id}

\item{region}{to look up}

\item{type}{one of "ranked" or "summary"}
}
\value{
summoner stats
}
\description{
Get ranked/summary stats by summoner ID
}
\examples{
get_stats(5908, "na", "ranked")
}

