% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summoner.R
\name{get_summoner}
\alias{get_summoner}
\title{Get summoner objects mapped by summoner ID for a given list of summoner IDs}
\usage{
get_summoner(ids, region = "na")
}
\arguments{
\item{ids}{list of summoner ids}

\item{region}{region to run query for}
}
\value{
list of summoner objects
}
\description{
Get summoner objects mapped by summoner ID for a given list of summoner IDs
}
\examples{
get_summoner(521955)
}

