% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summoner.R
\name{get_summoner_by_name}
\alias{get_summoner_by_name}
\title{Get summoner objects mapped by standardized summoner name for a given list of summoner names.}
\usage{
get_summoner_by_name(ids, region = "na")
}
\arguments{
\item{ids}{list of summoner ids}

\item{region}{region to run query for}
}
\value{
list of summoner objects
}
\description{
Get summoner objects mapped by standardized summoner name for a given list of summoner names.
}
\examples{
get_summoner_by_name("wildturtle")
}

