% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{get_summoner_games}
\alias{get_summoner_games}
\title{Get recent games by summoner ID}
\usage{
get_summoner_games(id, region = "na")
}
\arguments{
\item{id}{summoner id}

\item{region}{region to run query for}
}
\value{
data frame of recent games by summoner
}
\description{
Get recent games by summoner ID
}
\examples{
get_summoner_games(29400662)
}

