% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summoner.R
\name{get_summoner_masteries}
\alias{get_summoner_masteries}
\title{Get mastery pages mapped by summoner ID for a given list of summoner IDs}
\usage{
get_summoner_masteries(ids, region = "na")
}
\arguments{
\item{ids}{list of summoner ids}

\item{region}{region to run query for}
}
\value{
mastery information for given summoners
}
\description{
Get mastery pages mapped by summoner ID for a given list of summoner IDs
}
\examples{
get_summoner_masteries(c(29400662, 521955))
get_summoner_masteries("29400662")
}

