% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summoner.R
\name{get_summoner_name}
\alias{get_summoner_name}
\title{Get summoner names mapped by summoner ID for a given list of summoner IDs.}
\usage{
get_summoner_name(ids, region = "na")
}
\arguments{
\item{ids}{list of summoner ids}

\item{region}{region to run query for}
}
\value{
list of summoner names
}
\description{
Get summoner names mapped by summoner ID for a given list of summoner IDs.
}
\examples{
get_summoner_name(c("521955", 32110)) \%>\% names
}

