% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summoner.R
\name{get_summoner_runes}
\alias{get_summoner_runes}
\title{Get rune pages mapped by summoner ID for a given list of summoner IDs}
\usage{
get_summoner_runes(ids, region = "na")
}
\arguments{
\item{ids}{list of summoner ids}

\item{region}{region to run query for}
}
\value{
rune page info
}
\description{
Get rune pages mapped by summoner ID for a given list of summoner IDs
}
\examples{
get_summoner_runes("521955")
}

