% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riot_key.R
\name{set_riot_key}
\alias{set_riot_key}
\title{Set Riot API key}
\usage{
set_riot_key(key)
}
\arguments{
\item{key}{your Riot API key}
}
\value{
sets an environment variable
}
\description{
Set Riot API key
}
\examples{
\dontrun{set_riot_key("testkey-123")}
}

