% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{static_query}
\alias{static_query}
\title{Query LOL static API}
\usage{
static_query(region, type)
}
\arguments{
\item{region}{id}

\item{type}{see documentation https://developer.riotgames.com/api/methods#!/1055/3621}
}
\value{
url
}
\description{
Query LOL static API
}
\examples{
static_query("na", "item")
}

