% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_game.R
\name{get_current_game}
\alias{get_current_game}
\title{Get current game information for the given summoner ID}
\usage{
get_current_game(platform_id, summoner_id)
}
\arguments{
\item{summoner_id}{}
}
\description{
Get current game information for the given summoner ID
}

